SERVER/requirements.txt

appdirs==1.4.3
Django==4.0.2
django-extensions==3.1.1
djangorestframework==3.13.1
packaging==16.8
pyparsing==2.2.0
pytz==2021.3
sqlparse==0.2.2
asgiref==3.4.1
six==1.10.0

psycopg2==2.9.3
#Summary: psycopg2 - Python-PostgreSQL Database Adapter


